+++ mne/datasets/_fsaverage/base.py
@@ -7,8 +7,7 @@ import os.path as op
 
 
 from ..utils import _manifest_check_download, _get_path
-from ..._freesurfer import get_subjects_dir
-from ...utils import verbose, set_config
+from ...utils import verbose, set_config, get_subjects_dir
 
 FSAVERAGE_MANIFEST_PATH = op.dirname(__file__)
