+++ mne/label.py
@@ -13,7 +13,7 @@ import re
 
 import numpy as np
 
-from ._freesurfer import read_freesurfer_lut, _import_nibabel, get_subjects_dir
+from ._freesurfer import read_freesurfer_lut, _import_nibabel
 from .morph_map import read_morph_map
 from .parallel import parallel_func, check_n_jobs
 from .source_estimate import (SourceEstimate, VolSourceEstimate,
@@ -23,7 +23,7 @@ from .source_space import add_source_space_distances, SourceSpaces
 from .stats.cluster_level import _find_clusters, _get_components
 from .surface import read_surface, fast_cross_3d, mesh_edges, mesh_dist
 from .transforms import apply_trans
-from .utils import (_check_subject, logger, verbose, warn,
+from .utils import (_check_subject, logger, verbose, warn, get_subjects_dir,
                     check_random_state, _validate_type, fill_doc,
                     _check_option, check_version)
