+++ mne/source_space.py
@@ -30,13 +30,13 @@ from .surface import (read_surface, _create_surf_spacing, _get_ico_surface,
                       _normalize_vectors, _triangle_neighbors, mesh_dist,
                       complete_surface_info, _compute_nearest, fast_cross_3d,
                       _CheckInside)
-from ._freesurfer import (get_subjects_dir, _get_mri_info_data,
-                          _get_atlas_values)
+from ._freesurfer import (_get_mri_info_data, _get_atlas_values,
+                          read_freesurfer_lut)
 from .utils import (check_fname, logger, verbose, fill_doc,
                     _ensure_int, check_version, _get_call_line, warn,
                     _check_fname, _check_path_like, _check_sphere,
                     _validate_type, _check_option, _is_numeric, _pl, _suggest,
-                    object_size, sizeof_fmt, read_freesurfer_lut)
+                    object_size, sizeof_fmt, get_subjects_dir)
 from .parallel import parallel_func, check_n_jobs
 from .transforms import (invert_transform, apply_trans, _print_coord_trans,
                          combine_transforms, _get_trans,
